+++ src/scripts/lcoa_inputs_from_tlo_analyses/analysis_effect_of_treatment_ids.py
@@ -3,445 +3,62 @@
 import argparse
 import glob
 import os
-import textwrap
-import warnings
 import zipfile
 from pathlib import Path
-from typing import Tuple
 
 import numpy as np
-import pandas as pd
 from matplotlib import pyplot as plt
-from matplotlib.lines import Line2D
-from matplotlib.patches import Patch
-
-from scripts.calibration_analyses.analysis_scripts import plot_legends
-from scripts.lcoa_inputs_from_tlo_analyses.scenario_effect_of_treatment_ids import (
-    EffectOfEachTreatment,
-)
+import pandas as pd
 from tlo import Date
+from scripts.calibration_analyses.analysis_scripts import plot_legends
+from scripts.lcoa_inputs_from_tlo_analyses.fig_utils import (
+    do_bar_plot_with_ci,
+    plot_multiindex_dot_with_interval,
+)
+from scripts.lcoa_inputs_from_tlo_analyses.results_processing_utils import (
+    extract_deaths_total,
+    format_scenario_name,
+    get_counts_of_appts,
+    get_counts_of_hsi_by_short_treatment_id,
+    get_num_dalys_by_cause_label,
+    get_num_deaths_by_cause_label,
+    get_parameter_names_from_scenario_file,
+    get_periods_within_target_period,
+    get_total_num_dalys_by_agegrp_and_label,
+    get_total_num_death_by_agegrp_and_label,
+    get_total_population_by_year,
+    make_get_num_dalys_by_cause_label_and_period,
+    set_param_names_as_column_index_level_0,
+    target_period,
+)
+from scripts.costing.cost_estimation import (
+    apply_discounting_to_cost_data,
+    do_line_plot_of_cost,
+    do_stacked_bar_plot_of_cost_by_category,
+    estimate_input_cost_of_scenarios,
+    estimate_projected_health_spending,
+    extract_roi_at_specific_implementation_costs,
+    generate_multiple_scenarios_roi_plot,
+    load_unit_cost_assumptions,
+    summarize_cost_data,
+    tabulate_roi_estimates,
+)
 from tlo.analysis.utils import (
-    CAUSE_OF_DEATH_OR_DALY_LABEL_TO_COLOR_MAP,
+    compute_summary_statistics,
     extract_results,
-    get_color_cause_of_death_or_daly_label,
     get_color_short_treatment_id,
     make_age_grp_lookup,
-    make_age_grp_types,
-    order_of_cause_of_death_or_daly_label,
     squarify_neat,
     summarize,
-    to_age_group,
-    compute_summary_statistics,
+    unflatten_flattened_multi_index_in_logging,
 )
 
 TARGET_PERIOD = (Date(2026, 1, 1), Date(2041, 1, 1))
 PERIOD_LENGTH_YEARS_FOR_BAR_PLOTS = 5
 results_folder = Path("outputs/s.bhatia@imperial.ac.uk/effect_of_each_treatment_id-2026-02-16T154500Z")
-## retrieved from the suspended run in outputs/s.bhatia@imperial.ac.uk/effect_of_each_treatment_id-2026-02-12T120859Z
+# SCALING_FACTOR retrieved from the suspended run in
+# outputs/s.bhatia@imperial.ac.uk/effect_of_each_treatment_id-2026-02-12T120859Z
 SCALING_FACTOR = 58.158436
-
-
-def get_total_population_by_year(_df):
-    years_needed = [i.year for i in TARGET_PERIOD]
-    _df['year'] = pd.to_datetime(_df['date']).dt.year
-
-    # Filter for relevant years and return the total population as a Series
-    return \
-        _df.loc[_df['year'].between(min(years_needed), max(years_needed)), ['year', 'total']].set_index('year')[
-            'total']
-
-
-def extract_deaths_total(df: pd.DataFrame) -> pd.Series:
-    return pd.Series({"Total": len(df)})
-
-def target_period(target_period_tuple: tuple[Date, Date] = TARGET_PERIOD) -> str:
-    """Returns the target period as a string of the form YYYY-YYYY."""
-    return "-".join(str(t.year) for t in target_period_tuple)
-
-
-def get_periods_within_target_period(
-    period_length_years: int,
-    target_period_tuple: tuple[Date, Date] = TARGET_PERIOD,
-) -> list[tuple[str, tuple[int, int]]]:
-    """Return chunks within target period as [(label, (start_year, end_year)), ...]."""
-    if period_length_years <= 0:
-        raise ValueError("period_length_years must be a positive integer.")
-    start_year, end_year = target_period_tuple[0].year, target_period_tuple[1].year
-    periods = []
-    for chunk_start in range(start_year, end_year + 1, period_length_years):
-        chunk_end = min(chunk_start + period_length_years - 1, end_year)
-        periods.append((f"{chunk_start}-{chunk_end}", (chunk_start, chunk_end)))
-    return periods
-
-
-def get_parameter_names_from_scenario_file() -> Tuple[str]:
-    """Get tuple of scenario names from Scenario class used to create results."""
-    e = EffectOfEachTreatment()
-    return tuple(e._scenarios.keys())
-
-
-def format_scenario_name(_sn: str) -> str:
-    """Return reformatted scenario name ready for plotting."""
-    if _sn == "Nothing":
-        return "Nothing"
-    return _sn.lstrip("Only ")
-
-
-def set_param_names_as_column_index_level_0(_df: pd.DataFrame, param_names: tuple[str, ...]) -> pd.DataFrame:
-    """Set columns index level 0 as scenario param names."""
-    ordered_param_names_no_prefix = {i: x for i, x in enumerate(param_names)}
-    names_of_cols_level0 = [ordered_param_names_no_prefix.get(col) for col in _df.columns.levels[0]]
-    assert len(names_of_cols_level0) == len(_df.columns.levels[0])
-
-    reformatted_names = map(format_scenario_name, names_of_cols_level0)
-    _df.columns = _df.columns.set_levels(reformatted_names, level=0)
-    return _df
-
-
-def find_difference_extra_relative_to_comparison(
-    _ser: pd.Series,
-    comparison: str,
-    scaled: bool = False,
-    drop_comparison: bool = True,
-):
-    """Find run-wise differences relative to comparison in a series with multi-index."""
-    return (
-        _ser.unstack()
-        .apply(lambda x: (x - x[comparison]) / (x[comparison] if scaled else 1.0), axis=0)
-        .drop(index=([comparison] if drop_comparison else []))
-        .stack()
-    )
-
-
-def find_mean_difference_in_appts_relative_to_comparison(
-    _df: pd.DataFrame,
-    comparison: str,
-    drop_comparison: bool = True,
-):
-    """Find mean fewer appointments when treatment does not happen relative to comparison."""
-    return -summarize(
-        pd.concat(
-            {
-                _idx: find_difference_extra_relative_to_comparison(
-                    row, comparison=comparison, drop_comparison=drop_comparison
-                )
-                for _idx, row in _df.iterrows()
-            },
-            axis=1,
-        ).T,
-        only_mean=True,
-    )
-
-
-def find_mean_difference_extra_relative_to_comparison_dataframe(
-    _df: pd.DataFrame,
-    comparison: str,
-    drop_comparison: bool = True,
-):
-    """Same as find_difference_extra_relative_to_comparison but for dataframe."""
-    return summarize(
-        pd.concat(
-            {
-                _idx: find_difference_extra_relative_to_comparison(
-                    row, comparison=comparison, drop_comparison=drop_comparison
-                )
-                for _idx, row in _df.iterrows()
-            },
-            axis=1,
-        ).T,
-        only_mean=True,
-    )
-
-
-def get_num_deaths_by_cause_label(_df: pd.DataFrame, target_period_tuple: tuple[Date, Date] = TARGET_PERIOD) -> pd.Series:
-    """Return total deaths by label within target period."""
-    return _df.loc[pd.to_datetime(_df.date).between(*target_period_tuple)].groupby(_df["label"]).size()
-
-
-def get_num_dalys_by_cause_label(_df: pd.DataFrame, target_period_tuple: tuple[Date, Date] = TARGET_PERIOD) -> pd.Series:
-    """Return total DALYS by label within target period."""
-    return (
-        _df.loc[_df.year.between(*[i.year for i in target_period_tuple])]
-        .drop(columns=["date", "sex", "age_range", "year"])
-        .sum()
-    )
-
-
-def make_get_num_deaths_by_cause_label_and_period(
-    period_length_years: int,
-    target_period_tuple: tuple[Date, Date] = TARGET_PERIOD,
-):
-    """Create helper that summarizes deaths by cause and period chunks + overall."""
-    periods = get_periods_within_target_period(
-        period_length_years=period_length_years,
-        target_period_tuple=target_period_tuple,
-    )
-    period_lookup = {
-        year: period_label
-        for period_label, (start_year, end_year) in periods
-        for year in range(start_year, end_year + 1)
-    }
-    target_period_label = target_period(target_period_tuple)
-
-    def _get_num_deaths_by_cause_label_and_period(_df: pd.DataFrame) -> pd.Series:
-        _df_in_target = _df.loc[pd.to_datetime(_df.date).between(*target_period_tuple)].copy()
-        _df_in_target["year"] = pd.to_datetime(_df_in_target["date"]).dt.year
-        _df_in_target["period"] = _df_in_target["year"].map(period_lookup)
-
-        chunked = _df_in_target.groupby(["label", "period"]).size()
-        overall = _df_in_target.groupby("label").size()
-        overall.index = pd.MultiIndex.from_arrays(
-            [overall.index, np.repeat(target_period_label, len(overall.index))], names=["label", "period"]
-        )
-        return pd.concat([chunked, overall]).sort_index()
-
-    return _get_num_deaths_by_cause_label_and_period
-
-
-def make_get_num_dalys_by_cause_label_and_period(
-    period_length_years: int,
-    target_period_tuple: tuple[Date, Date] = TARGET_PERIOD,
-):
-    """Create helper that summarizes DALYS by cause and period chunks + overall."""
-    periods = get_periods_within_target_period(
-        period_length_years=period_length_years,
-        target_period_tuple=target_period_tuple,
-    )
-    period_lookup = {
-        year: period_label
-        for period_label, (period_start, period_end) in periods
-        for year in range(period_start, period_end + 1)
-    }
-    start_year, end_year = target_period_tuple[0].year, target_period_tuple[1].year
-    target_period_label = target_period(target_period_tuple)
-
-    def _get_num_dalys_by_cause_label_and_period(_df: pd.DataFrame) -> pd.Series:
-        _df_in_target = _df.loc[_df.year.between(start_year, end_year)].copy()
-        _df_in_target["period"] = _df_in_target["year"].map(period_lookup)
-
-        melted = (
-            _df_in_target.drop(columns=["date", "sex", "age_range"])
-            .melt(id_vars=["year", "period"], var_name="label", value_name="dalys")
-        )
-        chunked = melted.groupby(["label", "period"])["dalys"].sum()
-        overall = melted.groupby("label")["dalys"].sum()
-        overall.index = pd.MultiIndex.from_arrays(
-            [overall.index, np.repeat(target_period_label, len(overall.index))], names=["label", "period"]
-        )
-        return pd.concat([chunked, overall]).sort_index()
-
-    return _get_num_dalys_by_cause_label_and_period
-
-
-def do_bar_plot_with_ci(
-    _df: pd.DataFrame,
-    _param,
-    _ax,
-    period_labels_for_bar_plots: list[str],
-    target_period_label: str,
-):
-    """Make vertical bars by cause, decomposed into period chunks, with overall-period CI."""
-    available_params = (
-        _df.columns.get_level_values(0)
-        if isinstance(_df.columns, pd.MultiIndex)
-        else _df.columns
-    )
-    if _param not in available_params:
-        warnings.warn(f"Parameter '{_param}' not found in dataframe columns. Skipping plot.", stacklevel=2)
-        return
-
-    _df_nothing = _df[_param]
-    _df_nothing = _df_nothing.reindex(
-        pd.MultiIndex.from_product(
-            [CAUSE_OF_DEATH_OR_DALY_LABEL_TO_COLOR_MAP.keys(), period_labels_for_bar_plots + [target_period_label]],
-            names=["label", "period"],
-        ),
-        fill_value=0.0,
-    )
-    _df_nothing = _df_nothing.sort_index(axis=0, level=0, key=order_of_cause_of_death_or_daly_label)
-
-    cause_labels = list(_df_nothing.index.get_level_values("label").unique())
-
-    for i, cause_label in enumerate(cause_labels):
-        color = get_color_cause_of_death_or_daly_label(cause_label)
-        one_cause = _df_nothing.xs(cause_label, level="label")
-
-        bottom = 0.0
-        for j, period_label in enumerate(period_labels_for_bar_plots):
-            chunk_height = one_cause.loc[period_label, "mean"] if period_label in one_cause.index else 0.0
-            _ax.bar(i, chunk_height, bottom=bottom, color=color, alpha=0.9 if j % 2 == 0 else 0.35)
-            bottom += chunk_height
-
-        mean_value = one_cause.loc[target_period_label, "mean"]
-        lower_value = one_cause.loc[target_period_label, "lower"]
-        upper_value = one_cause.loc[target_period_label, "upper"]
-        overall_yerr = np.array([[mean_value - lower_value], [upper_value - mean_value]])
-        _ax.errorbar(i, mean_value, yerr=overall_yerr, fmt="none", ecolor="black", capsize=2, linewidth=1.2)
-
-    _ax.set_xticks(range(len(cause_labels)))
-    _ax.set_xticklabels(cause_labels, rotation=90)
-    chunk_legend_handles = [
-        Patch(facecolor="grey", alpha=0.9 if i % 2 == 0 else 0.35, label=period_label)
-        for i, period_label in enumerate(period_labels_for_bar_plots)
-    ]
-    ci_legend_handle = Line2D([0], [0], color="black", marker="|", markersize=8, linewidth=1.2, label="95% CI")
-    _ax.legend(handles=chunk_legend_handles + [ci_legend_handle], loc="upper right")
-
-
-def plot_multiindex_dot_with_interval(
-    _df: pd.DataFrame,
-    year: int,
-    _ax,
-    central_measure: str = "mean",
-    value_col: str = "population",
-    sort: bool = True,
-    x_label_rotation: int = 90,
-    x_tick_fontsize: int = 8,
-    label_wrap_width: int = 18,
-    max_xticks: int = 30,
-):
-    """Plot central-value dots and lower/upper intervals by category for one year."""
-    if not isinstance(_df.index, pd.MultiIndex) or _df.index.nlevels < 3:
-        raise ValueError("_df index must be a MultiIndex with at least 3 levels: category, stat, year.")
-    if value_col not in _df.columns:
-        raise ValueError(f"Column '{value_col}' not found in dataframe.")
-
-    year_level_values = _df.index.get_level_values(2)
-    available_years = pd.Index(year_level_values.unique()).sort_values()
-    if year not in available_years:
-        raise ValueError(f"Year '{year}' not found in index level 2. Available years: {available_years.tolist()}")
-
-    stat_level_values = _df.index.get_level_values(1)
-    required_stats = {central_measure, "lower", "upper"}
-    missing_stats = required_stats.difference(set(stat_level_values))
-    if missing_stats:
-        raise ValueError(
-            f"Missing required stat(s) in index level 1: {sorted(missing_stats)}. "
-            f"Available stats: {sorted(set(stat_level_values))}"
-        )
-
-    _plot = _df.xs(year, level=2)[value_col].unstack(level=1)
-    _plot = _plot.loc[:, [central_measure, "lower", "upper"]]
-    _plot = _plot.dropna(subset=[central_measure, "lower", "upper"])
-    if _plot.empty:
-        raise ValueError(f"No plottable rows remain for year '{year}' after selecting required stats.")
-
-    if sort:
-        _plot = _plot.sort_values(by=central_measure, ascending=True)
-
-    x = np.arange(len(_plot.index))
-    _ax.vlines(x, _plot["lower"], _plot["upper"], color="black", linewidth=1.2)
-    _ax.scatter(x, _plot[central_measure], color="black", s=20, zorder=3)
-
-    # Improve readability when category labels are long and/or numerous.
-    _ax.figure.set_size_inches(max(12, min(0.25 * len(_plot.index), 36)), 7)
-    wrapped_labels = [textwrap.fill(str(label), width=label_wrap_width) for label in _plot.index]
-    if max_xticks is not None and len(x) > max_xticks:
-        step = int(np.ceil(len(x) / max_xticks))
-        shown_positions = x[::step]
-        shown_labels = [wrapped_labels[i] for i in shown_positions]
-        _ax.set_xticks(shown_positions)
-        _ax.set_xticklabels(shown_labels, rotation=x_label_rotation, ha="right", fontsize=x_tick_fontsize)
-    else:
-        _ax.set_xticks(x)
-        _ax.set_xticklabels(wrapped_labels, rotation=x_label_rotation, ha="right", fontsize=x_tick_fontsize)
-    _ax.set_xlabel(_df.index.names[0] if _df.index.names[0] is not None else "category")
-    _ax.set_ylabel(value_col)
-    _ax.set_title(f"{value_col}: {central_measure} with lower/upper ({year})")
-    _ax.grid(axis="y")
-    _ax.spines["top"].set_visible(False)
-    _ax.spines["right"].set_visible(False)
-
-    return _ax
-
-
-def get_num_deaths_by_age_group(
-    _df: pd.DataFrame,
-    age_grp_lookup: dict,
-    target_period_tuple: tuple[Date, Date] = TARGET_PERIOD,
-):
-    """Return total deaths by age-group in target period."""
-    return (
-        _df.loc[pd.to_datetime(_df.date).between(*target_period_tuple)]
-        .groupby(_df["age"].map(age_grp_lookup).astype(make_age_grp_types()))
-        .size()
-    )
-
-
-def do_barh_plot_with_ci(_df: pd.DataFrame, _ax):
-    """Make horizontal bar plot for each treatment id."""
-    errors = pd.concat([_df["mean"] - _df["lower"], _df["upper"] - _df["mean"]], axis=1).T.to_numpy()
-    _df.plot.barh(ax=_ax, y="mean", xerr=errors, legend=False, color=[get_color_short_treatment_id(_id) for _id in _df.index])
-
-
-def do_label_barh_plot(_df: pd.DataFrame, _ax):
-    """Add text annotation from values in dataframe onto axis."""
-    y_cords = {ylabel.get_text(): ytick for ytick, ylabel in zip(_ax.get_yticks(), _ax.get_yticklabels())}
-    pos_on_rhs = _ax.get_xticks()[-1]
-
-    for label, row in _df.iterrows():
-        if row["mean"] > 0:
-            annotation = f"{round(row['mean'], 1)} ({round(row['lower'])}-{round(row['upper'])}) %"
-            _ax.annotate(
-                annotation,
-                xy=(pos_on_rhs, y_cords.get(label)),
-                xycoords="data",
-                horizontalalignment="left",
-                verticalalignment="center",
-                size=7,
-            )
-
-
-def get_total_num_death_by_agegrp_and_label(
-    _df: pd.DataFrame,
-    target_period_tuple: tuple[Date, Date] = TARGET_PERIOD,
-) -> pd.Series:
-    """Return deaths in target period by age-group and cause label."""
-    _df_limited_to_dates = _df.loc[_df["date"].between(*target_period_tuple)]
-    age_group = to_age_group(_df_limited_to_dates["age"])
-    return _df_limited_to_dates.groupby([age_group, "label"])["person_id"].size()
-
-
-def get_total_num_dalys_by_agegrp_and_label(
-    _df: pd.DataFrame,
-    target_period_tuple: tuple[Date, Date] = TARGET_PERIOD,
-) -> pd.Series:
-    """Return DALYS in target period by age-group and cause label."""
-    return (
-        _df.loc[_df.year.between(*[i.year for i in target_period_tuple])]
-        .assign(age_group=_df["age_range"])
-        .drop(columns=["date", "year", "sex", "age_range"])
-        .melt(id_vars=["age_group"], var_name="label", value_name="dalys")
-        .groupby(by=["age_group", "label"])["dalys"]
-        .sum()
-    )
-
-
-def get_counts_of_hsi_by_short_treatment_id(
-    _df: pd.DataFrame,
-    target_period_tuple: tuple[Date, Date] = TARGET_PERIOD,
-) -> pd.Series:
-    """Get counts of short treatment ids occurring in target period."""
-    mask = pd.to_datetime(_df["date"]).between(*target_period_tuple)
-    _counts_by_treatment_id = (
-        _df.loc[mask, "TREATMENT_ID"]
-        .apply(pd.Series)
-        .sum()
-        .astype(int)
-    )
-    ##_short_treatment_id = _counts_by_treatment_id.index.map(lambda x: x.split("_")[0] + "*")
-    ##return _counts_by_treatment_id.groupby(by=_short_treatment_id).sum()
-    return _counts_by_treatment_id
-
-
-def get_counts_of_appts(_df: pd.DataFrame, target_period_tuple: tuple[Date, Date] = TARGET_PERIOD) -> pd.Series:
-    """Get counts of appointments of each type being used in target period."""
-    return (
-        _df.loc[pd.to_datetime(_df["date"]).between(*target_period_tuple), "Number_By_Appt_Type_Code"]
-        .apply(pd.Series)
-        .sum()
-        .astype(int)
-    )
 
 
 def apply(results_folder: Path, output_folder: Path, resourcefilepath: Path = None):
@@ -459,6 +76,24 @@ def apply(results_folder: Path, output_folder: Path, resourcefilepath: Path = No
     target_period_label = target_period(TARGET_PERIOD)
 
     param_names = get_parameter_names_from_scenario_file()
+
+    # Costs calculation
+    alternative_discount_rates = [
+        {"discount_rate_cost": 0.03, "discount_rate_health": 0, "discounting_scenario": 'WHO-CHOICE (0.03,0)'},
+        {"discount_rate_cost": 0.03, "discount_rate_health": 0.03, "discounting_scenario": 'MAIN (0.03,0.03)'}
+    ]
+
+    for rates in alternative_discount_rates:
+        discount_rate_cost = rates["discount_rate_cost"]
+        discount_rate_health = rates["discount_rate_health"]
+        input_costs = estimate_input_cost_of_scenarios(
+                          results_folder,
+                          resourcefilepath,
+                          cost_only_used_staff=True,
+                          _discount_rate=discount_rate_cost,
+                          _metric="median",)
+
+
 
 
     # Get total population by year
